+++ src/main/java/database/DBSearch/DBSearch.java
@@ -1,4 +1,4 @@
-package database;
+package database.DBSearch;
 
 public interface DBSearch {
     void dbSearch();
